feat: Criação do resource tasks e implementação de users

+++ src/user/user.service.ts
@@ -1,6 +1,12 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { Prisma, User } from 'generated/prisma';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { CreateUserDto } from './dto/create-user.dto';
+import { Prisma, User } from 'generated/prisma';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
@@ -8,27 +14,68 @@ export class UserService {
   @Inject()
   private readonly prisma: PrismaService;
 
-  async createUser(data: Prisma.UserCreateInput): Promise<User> {
+  async createUser(data: Prisma.UserCreateInput): Promise<CreateUserDto> {
+    const hasUser = await this.findUserByEmail(data.email);
+
+    if (hasUser) throw new BadRequestException('Usuário já cadastrado');
+
     const hashPassword = await bcrypt.hash(data.password, 10);
 
-    return await this.prisma.user.create({
-      data: { ...data, password: hashPassword },
-    });
+    try {
+      const user = await this.prisma.user.create({
+        data: { ...data, password: hashPassword },
+      });
+
+      return this.buildUser(user);
+    } catch (error) {
+      throw new BadRequestException('Erro ao cadastrar usuário');
+    }
   }
 
-  async findUser(id: string): Promise<User | null> {
-    return this.prisma.user.findUnique({ where: { id } });
+  async findUserById(id: string): Promise<CreateUserDto> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+
+    if (!user) throw new BadRequestException('Usuário nao encontrado');
+
+    return this.buildUser(user);
+  }
+
+  async findUserByEmail(email: string): Promise<CreateUserDto | null> {
+    const user = await this.prisma.user.findUnique({ where: { email } });
+
+    if (!user) throw new NotFoundException('Usuário nao encontrado');
+
+    return this.buildUser(user);
   }
 
   async updateUser(params: {
     id: string;
     data: Prisma.UserUpdateInput;
-  }): Promise<User> {
+  }): Promise<CreateUserDto> {
     const { id, data } = params;
+
+    const user = await this.findUserById(id);
+
+    if (!user) throw new BadRequestException('Usuário nao encontrado');
+
     return await this.prisma.user.update({ where: { id }, data });
   }
 
-  async deleteUser(id: string): Promise<User> {
+  async deleteUser(id: string): Promise<CreateUserDto> {
+    const user = await this.findUserById(id);
+
+    if (!user) throw new BadRequestException('Usuário nao encontrado');
+
     return await this.prisma.user.delete({ where: { id } });
+  }
+
+  private buildUser(user: User): CreateUserDto {
+    return {
+      id: user.id,
+      name: user.name,
+      email: user.email,
+      password: user.password,
+      createdAt: user.createdAt,
+    };
   }
 }
